IGNITE-12525: code review fixes.
