+++ modules/spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -14,4 +14,4 @@
 # limitations under the License.
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.boot.autoconfigure.ignite.IgniteNodeAutoConfiguration
+org.apache.ignite.springframework.boot.autoconfigure.IgniteAutoConfiguration
