+++ modules/spring-boot-client-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -14,4 +14,4 @@
 # limitations under the License.
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.boot.autoconfigure.ignite.IgniteClientAutoConfiguration
+org.apache.ignite.springframework.boot.autoconfigure.IgniteClientAutoConfiguration
